+++ gatsby-config.js
@@ -78,13 +78,6 @@ const config = {
     {
       resolve: 'gatsby-source-filesystem',
       options: {
-        name: 'markdown-pages',
-        path: `${ __dirname }/src/markdown/`
-      }
-    },
-    {
-      resolve: 'gatsby-source-filesystem',
-      options: {
         name: 'images',
         path: `${ __dirname }/src/img`
       }
