+++ gatsby-ssr.js
@@ -5,6 +5,4 @@
  */
 
 // You can delete this file if you're not using it
-export { wrapRootElement } from './src/apollo/wrap-root-element';
-
-
+export { wrapRootElement } from './wrap-root-element'
